% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Run application locally}
\usage{
runApp()
}
\description{
Run application locally
}
\note{
A web version of the application is available at \url{https://pedometria.shinyapps.io/ptfs4br/}.
}
\examples{
\dontrun{
runApp()
}
}
